v1.0

+++ utilities/search.h
@@ -25,7 +25,7 @@ void search_lista(lista** list, int n) {
 
 void search_hash(hashNode** ht, int n) {
   auto tiempo = chrono_now();
-  for (size_t i = 0; i < n; i++) {
+  for (int i = 0; i < n; i++) {
     hashNode* nada = hash_search(ht, generar_numero());
   }
   chrono_get_time(tiempo);
@@ -33,7 +33,7 @@ void search_hash(hashNode** ht, int n) {
 
 void search_bst(avlNode **bst, int n) {
   auto tiempo = chrono_now();
-  for (size_t i = 0; i < n; i++) {
+  for (int i = 0; i < n; i++) {
     avlNode* nada = bst_search(bst, generar_numero());
   }
   chrono_get_time(tiempo);
